+++ hvpdb/__init__.py
@@ -1,2 +1,2 @@
 from .core import HVPDB as HVPDB
-__version__ = '1.0.4'
+__version__ = '1.0.4.post2'